Move name-related functions to new `data.name` module.  Small `data` edits.
Fixed some @classmethod function definitions, made a few methods
proper class functions.
